implemented Math commands

+++ src/command/math/ProductCommand.java
@@ -5,7 +5,8 @@ import command.Command;
 
 public class ProductCommand extends Command {
 
-        public double calculateValue(List<Object> param) {
-            return (double) Math.multiplyExact((long) param.get(0), (long) param.get(1));
-        }
+	public double calculateValue(List<Object> param) {
+		return (double) Math.multiplyExact((long) (double) param.get(0),
+				(long) (double) param.get(1));
+	}
 }
