+++ src/command/math/RemainderCommand.java
@@ -5,7 +5,8 @@ import command.Command;
 
 public class RemainderCommand extends Command {
 
-        public double calculateValue(List<Object> param) {
-            return (double) Math.floorMod((long) param.get(0), (long) param.get(1));
-        }
+	public double calculateValue(List<Object> param) {
+		return (double) Math.floorMod((long) (double) param.get(0),
+				(long) (double) param.get(1));
+	}
 }
